Moved responsibility canBeContacted to Neighbour
