+++ neighbour-discoverer/src/test/java/no/vegvesen/ixn/federation/discoverer/NeighbourDiscovererTest.java
@@ -161,34 +161,6 @@ public class NeighbourDiscovererTest {
 	}
 
 
-	/**
-	 * Tests for graceful backoff algorithm
-	 */
-
-	@Test
-	public void calculatedNextPostAttemptTimeIsInCorrectInterval(){
-		Neighbour ericsson = createNeighbour();
-		LocalDateTime now = LocalDateTime.now();
-
-		// Mocking the first backoff attempt, where the exponential is 0.
-		double exponential = 0;
-		long expectedBackoff = (long) Math.pow(2, exponential)*2; //
-
-		System.out.println("LocalDataTime now: "+ now.toString());
-		LocalDateTime lowerLimit = now.plusSeconds(expectedBackoff);
-		LocalDateTime upperLimit = now.plusSeconds(expectedBackoff+60);
-
-		System.out.println("Lower limit: " + lowerLimit.toString());
-		System.out.println("Upper limit: " + upperLimit.toString());
-
-		ericsson.setBackoffAttempts(0);
-		ericsson.setBackoffStart(now);
-
-		LocalDateTime result = neighbourDiscoverer.getNextPostAttemptTime(ericsson);
-
-		assertThat(result).isBetween(lowerLimit, upperLimit);
-	}
-
 	@Test
 	public void gracefulBackoffPostOfCapabilityDoesNotHappenBeforeAllowedPostTime(){
 		Neighbour ericsson = createNeighbour();
